too much nsm

+++ ft_strchr.c
@@ -6,7 +6,7 @@
 /*   By: krfranco <krfranco@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/07 16:44:59 by krfranco          #+#    #+#             */
-/*   Updated: 2023/11/08 15:42:18 by krfranco         ###   ########.fr       */
+/*   Updated: 2023/11/11 17:59:34 by krfranco         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strchr(const char *str, int c)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] || !c)
